started tests

+++ src/main/java/edu/austral/ingsis/math/Expression.java
@@ -1,6 +1,6 @@
 package edu.austral.ingsis.math;
 
 public interface Expression {
-    Double evaluate();
+    int evaluate();
     String toString();
 }
